test: relocate inline mod tests to tests/unit/
The previous coverage commits added #[cfg(test)] mod tests blocks
inside src/, which conflicts with the project convention of placing
tests under tests/unit/. Drop the inline blocks and re-add the cases
that can be expressed against the public API:

- ScreenRunner / ScreenRunnerContext non-tty error paths
- repo CLI commands (clear / list / play)
- OssInsightClient mock implementation

Tests that depended on private fields, private functions, or
internal mock plumbing are intentionally not ported here, since
the intent is to keep the public API the only test surface.
