+++ src/domain/repositories/challenge_repository.rs
@@ -347,112 +347,3 @@ impl ChallengeRepositoryInterface for ChallengeRepository {
         ChallengeRepository::list_cache_keys(self)
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::domain::models::loading::StepType;
-    use crate::infrastructure::storage::file_storage::FileStorage;
-    use crate::presentation::tui::screens::loading_screen::ProgressReporter;
-    use std::sync::atomic::{AtomicUsize, Ordering};
-
-    struct CountingProgressReporter {
-        calls: AtomicUsize,
-        processed: AtomicUsize,
-        total: AtomicUsize,
-    }
-
-    impl CountingProgressReporter {
-        fn new() -> Self {
-            Self {
-                calls: AtomicUsize::new(0),
-                processed: AtomicUsize::new(0),
-                total: AtomicUsize::new(0),
-            }
-        }
-    }
-
-    impl ProgressReporter for CountingProgressReporter {
-        fn set_step(&self, _step_type: StepType) {}
-
-        fn set_current_file(&self, _file: Option<String>) {}
-
-        fn set_file_counts(
-            &self,
-            step_type: StepType,
-            processed: usize,
-            total: usize,
-            current_file: Option<String>,
-        ) {
-            assert_eq!(step_type, StepType::CacheCheck);
-            assert!(current_file
-                .as_deref()
-                .is_some_and(|file| file.starts_with("Reconstructing challenge")));
-            self.calls.fetch_add(1, Ordering::SeqCst);
-            self.processed.store(processed, Ordering::SeqCst);
-            self.total.store(total, Ordering::SeqCst);
-        }
-    }
-
-    fn create_test_repo(root_path: PathBuf) -> GitRepository {
-        GitRepository {
-            user_name: "test".to_string(),
-            repository_name: "repo".to_string(),
-            remote_url: "https://github.com/test/repo".to_string(),
-            branch: Some("main".to_string()),
-            commit_hash: Some("reconstruct-source-file".to_string()),
-            is_dirty: false,
-            root_path: Some(root_path),
-        }
-    }
-
-    #[test]
-    fn save_then_load_reconstructs_challenge_from_source_file() {
-        let temp_dir = tempfile::tempdir().unwrap();
-        let repo_root = temp_dir.path().join("repo");
-        let source_dir = repo_root.join("src");
-        let source_path = source_dir.join("lib.rs");
-        let source_content =
-            "fn first() {}\nfn target() {\n    println!(\"cached\");\n}\nfn last() {}\n";
-        std::fs::create_dir_all(&source_dir).unwrap();
-        std::fs::write(&source_path, source_content).unwrap();
-
-        let mut file_storage = FileStorage::new();
-        file_storage.set_file_content(source_path.canonicalize().unwrap(), source_content.into());
-
-        let repository = ChallengeRepository {
-            cache_dir: temp_dir.path().join("cache"),
-            storage: Arc::new(CompressedFileStorage::new()),
-            file_storage: Arc::new(file_storage),
-        };
-        let git_repo = create_test_repo(repo_root);
-        let challenge =
-            Challenge::new("from-source".to_string(), "stale cache content".to_string())
-                .with_language("rust".to_string())
-                .with_difficulty_level(DifficultyLevel::Normal)
-                .with_source_info("src/lib.rs".to_string(), 2, 4)
-                .with_comment_ranges(vec![(3, 8)]);
-        let reporter = CountingProgressReporter::new();
-
-        repository.save_challenges(&git_repo, &[challenge]).unwrap();
-
-        let loaded = repository
-            .load_challenges_with_progress(&git_repo, Some(&reporter))
-            .unwrap();
-
-        assert_eq!(loaded.len(), 1);
-        assert_eq!(loaded[0].id, "from-source");
-        assert_eq!(
-            loaded[0].code_content,
-            "fn target() {\n    println!(\"cached\");\n}"
-        );
-        assert_eq!(loaded[0].source_file_path.as_deref(), Some("src/lib.rs"));
-        assert_eq!(loaded[0].start_line, Some(2));
-        assert_eq!(loaded[0].end_line, Some(4));
-        assert_eq!(loaded[0].comment_ranges, vec![(3, 8)]);
-        assert_eq!(loaded[0].difficulty_level, Some(DifficultyLevel::Normal));
-        assert_eq!(reporter.calls.load(Ordering::SeqCst), 1);
-        assert_eq!(reporter.processed.load(Ordering::SeqCst), 1);
-        assert_eq!(reporter.total.load(Ordering::SeqCst), 1);
-    }
-}
