+++ src/domain/repositories/version_repository.rs
@@ -146,60 +146,3 @@ impl VersionRepositoryInterface for VersionRepository {
         Box::pin(VersionRepository::fetch_latest_version(self))
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::infrastructure::http::github_api_client::GitHubApiClientFactoryImpl;
-    use chrono::Duration;
-
-    fn create_repository() -> VersionRepository {
-        VersionRepository {
-            github_client_factory: Arc::new(GitHubApiClientFactoryImpl::default()),
-            file_storage: Arc::new(FileStorage::new()),
-        }
-    }
-
-    fn create_cache_entry(last_checked: chrono::DateTime<Utc>) -> VersionCacheEntry {
-        VersionCacheEntry {
-            latest_version: "1.0.0".to_string(),
-            current_version: env!("CARGO_PKG_VERSION").to_string(),
-            update_available: false,
-            last_checked,
-        }
-    }
-
-    #[test]
-    fn is_cache_valid_accepts_recent_current_version_entry() {
-        let repository = create_repository();
-        let entry = create_cache_entry(Utc::now() - Duration::hours(1));
-
-        assert!(repository.is_cache_valid(&entry, 24));
-    }
-
-    #[test]
-    fn is_cache_valid_rejects_expired_entry() {
-        let repository = create_repository();
-        let entry = create_cache_entry(Utc::now() - Duration::hours(25));
-
-        assert!(!repository.is_cache_valid(&entry, 24));
-    }
-
-    #[test]
-    fn is_cache_valid_rejects_entry_for_different_current_version() {
-        let repository = create_repository();
-        let mut entry = create_cache_entry(Utc::now() - Duration::hours(1));
-        entry.current_version = "0.0.0".to_string();
-
-        assert!(!repository.is_cache_valid(&entry, 24));
-    }
-
-    #[test]
-    fn get_version_cache_path_uses_app_data_directory() {
-        let repository = create_repository();
-
-        let path = repository.get_version_cache_path().unwrap();
-
-        assert_eq!(path, PathBuf::from("/tmp/test/version_cache.json"));
-    }
-}
