+++ src/infrastructure/http/oss_insight_client.rs
@@ -187,34 +187,6 @@ mod mock_impl {
             OssInsightClient::fetch_trending_repositories(self, language, period).await
         }
     }
-
-    #[cfg(test)]
-    mod tests {
-        use super::*;
-
-        #[tokio::test]
-        async fn fetch_trending_repositories_returns_empty_mock_data() {
-            let client = OssInsightClient::new();
-
-            let repositories = client
-                .fetch_trending_repositories(Some("rust"), "weekly")
-                .await
-                .unwrap();
-
-            assert!(repositories.is_empty());
-        }
-
-        #[tokio::test]
-        async fn trait_fetch_trending_repositories_uses_mock_data() {
-            let client = OssInsightClient::new();
-            let repositories =
-                OssInsightClientInterface::fetch_trending_repositories(&client, None, "daily")
-                    .await
-                    .unwrap();
-
-            assert!(repositories.is_empty());
-        }
-    }
 }
 
 #[cfg(not(feature = "test-mocks"))]
