+++ src/presentation/cli/runner.rs
@@ -107,116 +107,3 @@ fn run_repo_command(repo_command: &RepoCommands) -> Result<()> {
         RepoCommands::Play => run_repo_play(),
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::run_cache_command;
-    use crate::domain::models::{Challenge, GitRepository};
-    use crate::domain::repositories::challenge_repository::ChallengeRepositoryInterface;
-    use crate::presentation::cli::args::CacheCommands;
-    use crate::presentation::tui::screens::loading_screen::ProgressReporter;
-    use crate::{GitTypeError, Result};
-
-    #[derive(Debug)]
-    struct MockChallengeRepository {
-        stats: std::result::Result<(usize, u64), String>,
-        keys: Vec<String>,
-    }
-
-    impl MockChallengeRepository {
-        fn with_stats(total_bytes: u64) -> Self {
-            Self {
-                stats: Ok((2, total_bytes)),
-                keys: vec![],
-            }
-        }
-
-        fn with_keys(keys: Vec<String>) -> Self {
-            Self {
-                stats: Ok((0, 0)),
-                keys,
-            }
-        }
-
-        fn with_stats_error(message: &str) -> Self {
-            Self {
-                stats: Err(message.to_string()),
-                keys: vec![],
-            }
-        }
-    }
-
-    impl ChallengeRepositoryInterface for MockChallengeRepository {
-        fn save_challenges(
-            &self,
-            _repo: &GitRepository,
-            _challenges: &[Challenge],
-            _reporter: Option<&dyn ProgressReporter>,
-        ) -> Result<()> {
-            Ok(())
-        }
-
-        fn load_challenges_with_progress(
-            &self,
-            _repo: &GitRepository,
-            _reporter: Option<&dyn ProgressReporter>,
-        ) -> Result<Option<Vec<Challenge>>> {
-            Ok(None)
-        }
-
-        fn get_cache_stats(&self) -> Result<(usize, u64)> {
-            self.stats
-                .clone()
-                .map_err(|message| GitTypeError::TerminalError(message.to_string()))
-        }
-
-        fn clear_cache(&self) -> Result<()> {
-            Ok(())
-        }
-
-        fn invalidate_repository(&self, _repo: &GitRepository) -> Result<bool> {
-            Ok(false)
-        }
-
-        fn list_cache_keys(&self) -> Result<Vec<String>> {
-            Ok(self.keys.clone())
-        }
-    }
-
-    #[test]
-    fn run_cache_stats_accepts_each_size_bucket() {
-        [512, 2 * 1024, 2 * 1024 * 1024, 2 * 1024 * 1024 * 1024]
-            .into_iter()
-            .map(MockChallengeRepository::with_stats)
-            .for_each(|repository| {
-                let result = run_cache_command(&CacheCommands::Stats, &repository);
-
-                assert!(result.is_ok());
-            });
-    }
-
-    #[test]
-    fn run_cache_list_accepts_existing_keys() {
-        let repository = MockChallengeRepository::with_keys(vec![
-            "github.com/owner/first:abc123".to_string(),
-            "github.com/owner/second:def456".to_string(),
-        ]);
-
-        let result = run_cache_command(&CacheCommands::List, &repository);
-
-        assert!(result.is_ok());
-    }
-
-    #[test]
-    fn run_cache_stats_wraps_repository_error() {
-        let repository = MockChallengeRepository::with_stats_error("stats unavailable");
-
-        let result = run_cache_command(&CacheCommands::Stats, &repository);
-
-        assert!(matches!(
-            result,
-            Err(GitTypeError::TerminalError(message))
-                if message == "Terminal error: stats unavailable"
-        ));
-    }
-}
