+++ src/presentation/cli/screen_runner.rs
@@ -191,87 +191,3 @@ impl Drop for ScreenRunnerContext {
         }
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::{run_screen, ScreenRunnerContext};
-    use crate::presentation::di::AppModule;
-    use crate::presentation::tui::screens::RepoListScreen;
-    use crate::presentation::tui::ScreenType;
-    use crate::GitTypeError;
-
-    fn assert_non_tty_terminal_error(result: crate::Result<impl Sized>) {
-        assert!(matches!(
-            result,
-            Err(GitTypeError::TerminalError(message))
-                if message.contains("Not running in a terminal environment")
-        ));
-    }
-
-    #[test]
-    fn run_screen_returns_terminal_error_without_tty() {
-        if atty::is(atty::Stream::Stdout) {
-            return;
-        }
-
-        let result = run_screen::<RepoListScreen, (), (), fn(&RepoListScreen) -> Option<()>>(
-            ScreenType::RepoList,
-            None::<()>,
-            None,
-        );
-
-        assert_non_tty_terminal_error(result);
-    }
-
-    #[test]
-    fn screen_runner_context_new_returns_terminal_error_without_tty() {
-        if atty::is(atty::Stream::Stdout) {
-            return;
-        }
-
-        let result = ScreenRunnerContext::new();
-
-        assert_non_tty_terminal_error(result);
-    }
-
-    #[test]
-    fn inactive_screen_runner_context_cleanup_is_noop() {
-        let context = ScreenRunnerContext {
-            container: AppModule::builder().build(),
-            terminal_active: false,
-        };
-
-        assert!(context.cleanup().is_ok());
-    }
-
-    #[test]
-    fn active_screen_runner_context_cleanup_runs_terminal_cleanup() {
-        let context = ScreenRunnerContext {
-            container: AppModule::builder().build(),
-            terminal_active: true,
-        };
-
-        assert!(context.cleanup().is_ok());
-    }
-
-    #[test]
-    fn screen_runner_context_run_screen_returns_terminal_error_without_tty() {
-        if atty::is(atty::Stream::Stdout) {
-            return;
-        }
-
-        let context = ScreenRunnerContext {
-            container: AppModule::builder().build(),
-            terminal_active: false,
-        };
-
-        let result = context
-            .run_screen::<RepoListScreen, (), (), fn(&RepoListScreen) -> Option<()>>(
-                ScreenType::RepoList,
-                None::<()>,
-                None,
-            );
-
-        assert_non_tty_terminal_error(result);
-    }
-}
