+++ src/presentation/tui/screen_manager.rs
@@ -946,57 +946,3 @@ impl ScreenManagerFactory for ScreenManagerFactoryImpl {
         manager
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use ratatui::backend::TestBackend;
-
-    fn create_test_manager(event_bus: Arc<EventBus>) -> ScreenManagerImpl<TestBackend> {
-        let challenge_store =
-            Arc::new(ChallengeStore::default()) as Arc<dyn ChallengeStoreInterface>;
-        let repository_store =
-            Arc::new(RepositoryStore::default()) as Arc<dyn RepositoryStoreInterface>;
-        let session_store = Arc::new(SessionStore::default()) as Arc<dyn SessionStoreInterface>;
-        let stage_repository = Arc::new(StageRepository::new(
-            None,
-            challenge_store.clone(),
-            repository_store,
-            session_store.clone(),
-        )) as Arc<dyn StageRepositoryInterface>;
-        let session_tracker: Arc<dyn SessionTrackerInterface> = Arc::new(SessionTracker::default());
-        let total_tracker: Arc<dyn TotalTrackerInterface> = Arc::new(TotalTracker::default());
-        let session_manager = Arc::new(SessionManager::new_with_dependencies(
-            event_bus.clone(),
-            stage_repository.clone(),
-            session_tracker,
-            total_tracker,
-        )) as Arc<dyn SessionManagerInterface>;
-        let terminal = Terminal::new(TestBackend::new(80, 24)).unwrap();
-
-        ScreenManagerImpl::new(
-            event_bus,
-            session_store,
-            session_manager,
-            stage_repository,
-            terminal,
-        )
-    }
-
-    #[test]
-    fn setup_event_subscriptions_stores_navigate_transition() {
-        let event_bus = Arc::new(EventBus::new());
-        let manager = Arc::new(Mutex::new(create_test_manager(event_bus.clone())));
-
-        ScreenManagerImpl::setup_event_subscriptions(&manager);
-        event_bus.publish(NavigateTo::Replace(ScreenType::Help));
-
-        let manager = manager.lock().unwrap();
-        let pending = manager.pending_transition.lock().unwrap().clone();
-
-        assert!(matches!(
-            pending,
-            Some(ScreenTransition::Replace(ScreenType::Help))
-        ));
-    }
-}
