+++ tests/unit/infrastructure/mod.rs
@@ -2,5 +2,6 @@ pub mod console_tests;
 pub mod database;
 pub mod git;
 pub mod logging_tests;
+pub mod oss_insight_client_tests;
 pub mod storage;
 pub mod terminal_tests;
