+++ tests/unit/presentation/mod.rs
@@ -1,5 +1,7 @@
 pub mod cli_placeholder_command_tests;
+pub mod cli_repo_command_tests;
 pub mod cli_runner_tests;
+pub mod cli_screen_runner_tests;
 pub mod cli_trending_tests;
 pub mod game;
 pub mod sharing_tests;
